requirements.txt
Django==4.2.7
Pillow==10.1.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-dotenv==1.0.0
weasyprint==60.1

# WeasyPrint requires GTK3 runtime
# Download and install from: https://github.com/tschoonj/GTK-for-Windows-Runtime-Environment-Installer/releases
